* IModel.cs: * IAdress.cs: * ICompany.cs: * IProduct.cs: Moved to folder and namespace Products.
* ahbsd.lib.csproj: Added folder and namespace Products
