+++ ahbsd.lib/Products/IModel.cs
@@ -14,7 +14,7 @@
 //    limitations under the License.
 using System.ComponentModel;
 
-namespace ahbsd.lib
+namespace ahbsd.lib.Products
 {
     /// <summary>
     /// Interface of an model of a product.
